Clean up clients (#8)
cleans up the clients directory
